grpclog: move component depth functions from loggerv2.go to component.go
